Revert "chore: Move middleware to folder and add tests (#1436)"
This reverts commit 3b169a19e58a7c7c916448db82c9063bf4f7ac86.
