--- lib/middleware/__test__/notificationsMiddleware.test.ts
@@ -1,9 +1,0 @@
-const { notifyAboutUpdates } = require('../notificationsMiddleware');
-
-describe('lib/middleware/notificationsMiddleware', () => {
-  describe('notifyAboutUpdates()', () => {
-    it('should safely execute without throwing an error', () => {
-      expect(() => notifyAboutUpdates()).not.toThrow();
-    });
-  });
-});
